requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
anndata==0.10.5.post1
annotated-types==0.6.0
antlr-denter==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_api_compat==1.4.1
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
automata-lib==8.1.0
Babel==2.14.0
backoff==2.2.1
bbknn==1.6.0
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.20.0
boto3==1.34.29
botocore==1.34.29
cached-property==1.5.2
cached_method==0.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.85
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contextlib2==21.6.0
contourpy==1.2.0
croniter==1.4.1
cycler==0.12.1
Cython==3.0.5
dateutils==0.6.12
debugpy==1.8.0
decorator==5.1.1
deepdiff==6.7.1
defusedxml==0.7.1
dm-tree==0.1.8
docker-pycreds==0.4.0
docrep==0.3.2
editor==1.6.6
etils==1.6.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
faiss-gpu==1.7.2
fastapi==0.109.0
fastjsonschema==2.19.1
fbpca==1.0
filelock==3.13.1
flax==0.8.1
fonttools==4.45.0
fqdn==1.5.1
frozendict==2.3.9
frozenlist==1.4.1
fsspec==2024.2.0
gensim==4.3.2
geosketch==1.2
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.26.1
google-auth-oauthlib==1.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
harmony==1.2.3360
harmonypy==0.0.9
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.6
igraph==0.10.8
importlib-resources==6.1.1
importlib_metadata==8.4.0
inquirer==3.2.3
intervaltree==3.1.0
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.31
jax-cuda12-pjrt==0.4.31
jax-cuda12-plugin==0.4.31
jaxlib==0.4.31
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
leidenalg==0.10.1
lightly==1.5.10
lightly-utils==0.0.2
lightning==2.1.4
lightning-cloud==0.5.62
lightning-utilities==0.10.1
llvmlite==0.41.1
loompy==3.0.7
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
ml_collections==0.1.1
mpmath==1.3.0
msgpack==1.0.7
mudata==0.2.3
multidict==6.0.5
multipledispatch==1.0.0
natsort==8.4.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.0.7
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.4
numpy-groupies==0.11.2
numpyro==0.13.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
optax==0.1.9
orbax-checkpoint==0.5.3
ordered-set==4.1.0
overrides==7.6.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
plottable==0.1.5
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydot==1.4.2
Pygments==2.17.2
PyJWT==2.8.0
pynndescent==0.5.11
pyparsing==3.1.1
pyro-api==0.1.2
pyro-ppl==1.8.6
pyspark==3.3.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==2.2.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
readchar==4.0.5
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
runs==1.2.2
s3transfer==0.10.0
scanorama==1.7.4
scanpy==1.9.6
scib-metrics==0.5.1
scikit-dimension==0.3.3
scikit-learn==1.4.0
scipy==1.12.0
scvi-tools==1.1.0.post2
seaborn==0.13.2
Send2Trash==1.8.2
sentry-sdk==1.39.1
session_info==1.0.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sparse==0.15.1
stack-data==0.6.3
starlette==0.35.1
starsessions==1.3.0
statsmodels==0.14.0
stdlib-list==0.10.0
style==1.1.0
submitit==1.5.1
symphonypy==0.2.1
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorstore==0.1.53
terminado==0.18.0
texttable==1.7.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.1
torch==2.4.0
torchmetrics==1.3.1
torchvision==0.19.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
triton==3.0.0
tweet-preprocessor==0.6.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
umap-learn==0.5.5
update==0.0.1
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.27.0.post1
wandb==0.16.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wordcloud==1.9.2
xarray==2024.1.1
xmod==1.8.1
yarl==1.9.4
zipp==3.17.0
